requirements.txt
Django
Pillow
environs
django-ckeditor
django-crispy_forms
